DBService

+++ src/album/album.controller.ts
@@ -4,42 +4,55 @@ import {
   Delete,
   Get,
   HttpCode,
+  HttpException,
+  HttpStatus,
   Param,
   Post,
   Put,
+  UsePipes,
+  ValidationPipe,
 } from '@nestjs/common';
+import { validate as uuidValidate } from 'uuid';
+import { Album } from 'src/interfaces/album.interface';
 import { AlbumService } from './album.service';
+import { CreateAlbumDto } from './dto/create-album.dto';
+import { UpdateAlbumDto } from './dto/update-album.dto';
 
 @Controller('/album')
 export class AlbumController {
   constructor(private readonly albumService: AlbumService) {}
 
   @Get()
-  getAlbums(): string {
+  getAlbums(): Album[] {
     return this.albumService.getAlbums();
   }
 
   @Get(':id')
-  // @HttpCode(200)
-  getAlbum(@Param('id') id) {
+  getAlbum(@Param('id') id: string): Album {
+    if (!uuidValidate(id))
+      throw new HttpException('Id is not UUID', HttpStatus.BAD_REQUEST);
     return this.albumService.getAlbum(id);
   }
 
+  @UsePipes(new ValidationPipe())
   @Post()
-  // @HttpCode(201)
-  createAlbum(@Body() album) {
+  createAlbum(@Body() album: CreateAlbumDto): Album {
     return this.albumService.createAlbum(album);
   }
 
+  @UsePipes(new ValidationPipe())
   @Put(':id')
-  // @HttpCode(200)
-  updateAlbum(@Param('id') id, @Body() album) {
+  updateAlbum(@Param('id') id: string, @Body() album: UpdateAlbumDto): Album {
+    if (!uuidValidate(id))
+      throw new HttpException('Id is not UUID', HttpStatus.BAD_REQUEST);
     return this.albumService.updateAlbum(id, album);
   }
 
   @Delete(':id')
   @HttpCode(204)
-  deleteAlbum(@Param('id') id) {
-    return this.albumService.deleteAlbum(id);
+  deleteAlbum(@Param('id') id: string): void {
+    if (!uuidValidate(id))
+      throw new HttpException('Id is not UUID', HttpStatus.BAD_REQUEST);
+    this.albumService.deleteAlbum(id);
   }
 }
