+++ src/favorites/favorites.controller.ts
@@ -1,4 +1,15 @@
-import { Controller, Delete, Get, HttpCode, Param, Post } from '@nestjs/common';
+import {
+  Controller,
+  Delete,
+  Get,
+  HttpCode,
+  HttpException,
+  HttpStatus,
+  Param,
+  Post,
+} from '@nestjs/common';
+import { validate as uuidValidate } from 'uuid';
+import { FavoritesRepsonse } from 'src/interfaces/favorites.interface';
 import { FavoritesService } from './favorites.service';
 
 @Controller('/favs')
@@ -6,43 +17,52 @@ export class FavoritesController {
   constructor(private readonly favoritesService: FavoritesService) {}
 
   @Get()
-  getFavorites(): string {
+  getFavorites(): FavoritesRepsonse {
     return this.favoritesService.getFavorites();
   }
 
   @Post('/track/:id')
-  // @HttpCode(201)
-  addTrack(@Param('id') id) {
+  addTrack(@Param('id') id: string) {
+    if (!uuidValidate(id))
+      throw new HttpException('Id is not UUID', HttpStatus.BAD_REQUEST);
     return this.favoritesService.addTrack(id);
   }
 
   @Delete('/track/:id')
   @HttpCode(204)
-  deleteTrack(@Param('id') id) {
-    return this.favoritesService.deleteTrack(id);
+  deleteTrack(@Param('id') id: string): void {
+    if (!uuidValidate(id))
+      throw new HttpException('Id is not UUID', HttpStatus.BAD_REQUEST);
+    this.favoritesService.deleteTrack(id);
   }
 
   @Post('/album/:id')
-  // @HttpCode(201)
-  addAlbum(@Param('id') id) {
+  addAlbum(@Param('id') id: string) {
+    if (!uuidValidate(id))
+      throw new HttpException('Id is not UUID', HttpStatus.BAD_REQUEST);
     return this.favoritesService.addAlbum(id);
   }
 
   @Delete('/album/:id')
   @HttpCode(204)
-  deleteAlbum(@Param('id') id) {
-    return this.favoritesService.deleteAlbum(id);
+  deleteAlbum(@Param('id') id: string): void {
+    if (!uuidValidate(id))
+      throw new HttpException('Id is not UUID', HttpStatus.BAD_REQUEST);
+    this.favoritesService.deleteAlbum(id);
   }
 
   @Post('/artist/:id')
-  // @HttpCode(201)
-  addArtist(@Param('id') id) {
+  addArtist(@Param('id') id: string) {
+    if (!uuidValidate(id))
+      throw new HttpException('Id is not UUID', HttpStatus.BAD_REQUEST);
     return this.favoritesService.addArtist(id);
   }
 
   @Delete('/artist/:id')
   @HttpCode(204)
-  deleteArtist(@Param('id') id) {
+  deleteArtist(@Param('id') id: string): void {
+    if (!uuidValidate(id))
+      throw new HttpException('Id is not UUID', HttpStatus.BAD_REQUEST);
     return this.favoritesService.deleteArtist(id);
   }
 }
